Improve backstory editor
