+++ src/main/python/plotlyst/view/widget/template/api.py
@@ -24,7 +24,7 @@ from overrides import overrides
 from src.main.python.plotlyst.core.domain import Character
 from src.main.python.plotlyst.core.template import TemplateField, SelectionItem, \
     enneagram_field, traits_field, ProfileTemplate
-from src.main.python.plotlyst.view.widget.character.control import EnneagramSelector
+from src.main.python.plotlyst.view.widget.character.editor import EnneagramSelector
 from src.main.python.plotlyst.view.widget.template.base import TemplateWidgetBase
 from src.main.python.plotlyst.view.widget.template.impl import TraitSelectionWidget
 from src.main.python.plotlyst.view.widget.template.profile import ProfileTemplateView
