+++ ui/character_editor.ui
@@ -794,7 +794,7 @@ border: 0px;
   <customwidget>
    <class>CharacterTimelineWidget</class>
    <extends>QWidget</extends>
-   <header>src.main.python.plotlyst.view.widget.characters</header>
+   <header>src.main.python.plotlyst.view.widget.character.editor</header>
    <container>1</container>
   </customwidget>
  </customwidgets>
